LPS-43148 Apply the same exception-handling pattern to UploadImageAction we have in EFEA
We should not add any exception to SessionErrors, as the ADD_TEMP action is an AJAX-request, thus, there won't be "render" phase executed for the "image_uploader" portlet where the errors map could be cleared.
This could lead to display UI error messages from a previous unsuccessfull action upon hiding and reopening the selector dialog.